requirements.txt
streamlit
pillow
pandas
boto3

